pyproject.toml
[tool.poetry]
name = "doc_data"
version = "0.3.2"
description = "Data collector for my PhD"
authors = ["Caio Geraldes <caioaguida@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "1.10.0"
numpy = "^1.22.1"
pandas = "^1.3.5"
scikit-learn = ">=0.24.2,<0.25.0"
scipy = "1.6.1"
beautifulsoup4 = "^4.10.0"
lxml = "^4.7.1"
cltk = "^1.0.22"
stanza = "^1.3.0"
coverage = "^6.3.1"
jupyter = "^1.0.0"
ipython = "^8.0.0"
notebook = "^6.4.7"
jupyterlab = "^3.2.8"
tqdm = "^4.62.3"
importlib-metadata= "^4.8.2"
python-dotenv = "^0.19.2"
pymongo = "^4.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.12b0"
pylint = "^2.12.2"
mypy = "^0.931"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

